# Optimize expensive XPath expressions in HTML transform