--- a/.eslintrc.json
+++ b/.eslintrc.json
@@ -101,6 +101,10 @@
                 "@typescript-eslint/no-unused-vars": "error",
                 "no-empty-function": "off",
                 "@typescript-eslint/no-empty-function": "error",
+                "no-undef": "off",
+                "no-redeclare": "off",
+                "@typescript-eslint/no-redeclare": "error",
+                "import/no-unresolved": "off",
 
                 "import/extensions": [
                     "error",