--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -23,6 +23,14 @@
         "editor.formatOnSave": true
     },
 
+    // Note: it would be nice to reverse this! But it's currently preventing a *ton* of whitespace diff noise
+    "[xml]": {
+        "editor.formatOnSave": false
+    },
+    "[xsl]": {
+        "editor.formatOnSave": false
+    },
+
     // Code navigation
     "javascript.referencesCodeLens.enabled": true
 }