--- a/package.json
+++ b/package.json
@@ -31,6 +31,7 @@
     "scripts": {
         "benchmarks": "vite-node ./test/benchmarks.ts",
         "build": "vite build && npm run tsc",
+        "prepare": "npm run build",
         "start": "vite",
         "eslint-check": "eslint app.ts src/**/*.ts vite.config.ts test/**/*.ts",
         "eslint-fix": "eslint app.ts src/**/*.ts vite.config.ts test/**/*.ts --fix",