--- a/typings/@prettier/plugin-xml.d.ts
+++ b/typings/@prettier/plugin-xml.d.ts
@@ -1,0 +1,20 @@
+declare module '@prettier/plugin-xml' {
+    import type { Plugin as PrettierPlugin, SupportLanguage } from 'prettier';
+
+    export interface PrettierPluginXMLOptions {
+        xmlSelfClosingSpace?: boolean;
+        xmlWhitespaceSensitivity?: 'ignore' | 'strict';
+    }
+
+    interface PrettierPluginXMLSupportLanguage extends SupportLanguage {
+        parsers: ['xml'];
+    }
+
+    export interface PrettierPluginXML extends PrettierPlugin {
+        languages: PrettierPluginXMLSupportLanguage;
+    }
+
+    declare const prettierPluginXML: PrettierPluginXML;
+
+    export default prettierPluginXML;
+}