… .eslintignore
… test/external-fixtures/enketo-core/SOAR_facilitysurvey_full_2017May11_Kenya.xml
… test/external-fixtures/enketo-core/autocomplete-cascade.xml
… test/external-fixtures/enketo-core/calc-control.xml
… test/external-fixtures/enketo-core/calc-in-group-with-relevant.xml
… test/external-fixtures/enketo-core/calcs-cascade.xml
… test/external-fixtures/enketo-core/calcs.xml
… test/external-fixtures/enketo-core/calcs_in_repeats.xml
… test/external-fixtures/enketo-core/calcs_in_repeats_2.xml
… test/external-fixtures/enketo-core/cascading_mixture_itext_noitext.xml
… test/external-fixtures/enketo-core/comment.xml
… test/external-fixtures/enketo-core/constraint-dependency.xml
… test/external-fixtures/enketo-core/count-repeated-nodes.xml
… test/external-fixtures/enketo-core/dynamic-required.xml
… test/external-fixtures/enketo-core/exclude-non-relevant-basic.xml
… test/external-fixtures/enketo-core/exclude-non-relevant-cascade-across-repeats.xml
… test/external-fixtures/enketo-core/exclude-non-relevant-repeat-calculate-single.xml
… test/external-fixtures/enketo-core/exclude-non-relevant-repeat-count.xml
… test/external-fixtures/enketo-core/exclude-non-relevant-repeat-irrelevant-date.xml
… test/external-fixtures/enketo-core/external-data-v2.xml
… test/external-fixtures/enketo-core/group-relevant-indexed-repeat.xml
… test/external-fixtures/enketo-core/group_branch.xml
… test/external-fixtures/enketo-core/groups-pages.xml
… test/external-fixtures/enketo-core/issue208.xml
… test/external-fixtures/enketo-core/itemset-relevant-2.xml
… test/external-fixtures/enketo-core/itemset-relevant.xml
… test/external-fixtures/enketo-core/itemset-repeat-relevant-current-new.xml
… test/external-fixtures/enketo-core/itemset-repeat-relevant-current-old.xml
… test/external-fixtures/enketo-core/jr-choice-name-autocomplete.xml
… test/external-fixtures/enketo-core/jr-choice-name-external.xml