After opening #164, it irked me that I knew of a [particularly impactful](https://github.com/enketo/enketo-transformer/pull/164#discussion_r1089338050) optimization candidate and hadn't addressed it. So this addresses it, as well as several other candidates which could similarly improve performance on forms which transform slowly. The results astounded me, so much so that we've decided to take a different approach to #164. But first, about this change.

### Substantive change

The improvement comes from using `xsl:key` elements to reference the expressions identified as expensive, and to look them up with the corresponding `key` XPath function. My strong assumption is that the reason this has such a big impact is that the non-key references are rerun on each visit to their respective `xsl:template` or upon each reference to a top-level `xsl:variable`, whereas `key` presumably pre-caches the expression's nodeset and only filters based on the `use` argument.

Here's what that looks like:

[Before](https://github.com/eyelidlessness/enketo-transformer/actions/runs/4108460351#summary-11150509210):

<table role="table">
<tbody><tr>
<th></th>
<th>Name</th>
<th>Result</th>
</tr>
<tr>
<td>Slowest</td>
<td>va_who_v1_5_3.xml</td>
<td>6444.79 ms (0.16 ops/s ±2.37%, 5 runs)</td>
</tr>
<tr>
<td>Fastest</td>
<td>form-id-with-accent.xml</td>
<td>9.70 ms (103 ops/s ±3.47%, 7 runs)</td>
</tr>
<tr>
<td>Average overall</td>
<td colspan="2">67.41ms</td>
</tr>
<tr>
<td>Average without outliers</td>
<td colspan="2">12.93ms</td>
</tr>
<tr>
<td>Total runtime</td>
<td colspan="2">628.08s</td>
</tr>
</tbody></table>

[After](https://github.com/eyelidlessness/enketo-transformer/actions/runs/4108594786#summary-11150891861).

<table role="table">
<tbody><tr>
<th></th>
<th>Name</th>
<th>Result</th>
</tr>
<tr>
<td>Slowest</td>
<td>va_who_v1_5_3.xml</td>
<td>543.47 ms (1.84 ops/s ±2.33%, 7 runs)</td>
</tr>
<tr>
<td>Fastest</td>
<td>arbitrary-html.xml</td>
<td>9.31 ms (107 ops/s ±1.29%, 7 runs)</td>
</tr>
<tr>
<td>Average overall</td>
<td colspan="2">18.59ms</td>
</tr>
<tr>
<td>Average without outliers</td>
<td colspan="2">12.35ms</td>
</tr>
<tr>
<td>Total runtime</td>
<td colspan="2">601.28s</td>
</tr>
</tbody></table>

I've seen similar or better improvement in my local dev environment—as with #164, the baseline performance is substantially worse but the change performs roughly the same as CI.

### The rest of the change

The change also incorporates:

- Snapshot tests originally introduced in #164 to catch potential regressions, as well as adding tests specifically for one part of the transform I found to be a noop. 
- A reworked benchmark suite using `benchmark.js`, because the ad-hoc one in #164 wasn't sitting well with me.

### What's next

These changes have a similar impact on the browser implementation in #164, but they largely eliminate any benefit of running a headless browser for the Node implementation. Besides that, the headless browser implementation adds quite a bit of unnecessary complexity.

We don't want to give up on the other present and potential future benefits discussed in that PR. So after some discussion with @lognaturel, we concluded that it's probably best instead to rework the changes in #164. That will entail:

1. A compatibility interface for the non-XSL portions of the transform, handling both DOM and `libxmljs` operations.
2. The same XSL to TypeScript changes, adapted to use that compatibility interface.
3. Adapt **all tests and benchmarks** to run in all target environments. (We should hopefully be able to run these concurrently in CI.)
4. Restrict headless browser usage to tests. (I expect that this will be the most sensible way to achieve the multi-environment test goal.)