@lognaturel 2023-02-08
@MartijnR eager to hear whether you/folks you talk to experience these perf benefits and how they manifest. We're always interested in more benchmark forms if you come across them.

Did I get https://github.com/enketo/enketo-transformer/pull/165#discussion_r1098036785 right? This has been removed because it [didn't work](https://github.com/enketo/enketo-transformer/pull/165#discussion_r1098876917).