@MartijnR 2023-02-08
I will let OpenClinica know about this major change. It looks like it may reduce Enketo's server requirements (to nothing!!).

I am also waiting to see transformation timing results on http://performance.enketo.org/ (User Graphs -> Graphite), but I am not sure if this app still updates successfully (probably doesn't). I no longer seem to have ssh access. Have you folks worked on this server perhaps? If not, I'll force access via the Digital Ocean droplet settings. 

Will check the comment and respond there!