@lognaturel 2023-02-08
You gave me SSH access last July with email instructions. I see at least two public keys there that belong to you.

We intended to do something with it but have not had the time. I can see that it has pulled the latest Core. I hadn't seen the transformer graph before, will be interesting to see if those forms show a notable improvement.