@MartijnR 2023-02-08
> I see at least two public keys there that belong to you.

Ah thanks, that will help figure out what is going on. 

> will be interesting to see if those forms show a notable improvement

Yes, especially the 250,000 ms transformation!