@lognaturel 2023-02-08
I unfortunately have to leave things in a bad state with dependency update failures. All this really leads me to wonder what the benchmarks were being run against previously. Will try to come back to it tonight unless someone else can take a look before.