@MartijnR 2023-02-08
I think it may be churning away again (but will have run an old version for quite a while before today). 

Fwiw:
I installed Node 14 (even though that shouldn't have been necessary). Removing node 16 was a little challenging (there seemed to be 2 versions running) so I ended up deleting the folder with the binary after which I could install 14 from nodesource.