@lognaturel 2023-02-08
Great, it does look like it's running. And yes, looks like it was running a really old version previously so we can't really meaningfully compare this latest set of improvements.

The worst transform time I'm seeing is 7647 for bench 6. Does that look right to you?