@lognaturel 2023-02-08
Have checked out core 6.1.7 and am doing a run on that. The graphs are messed up anyway so I don't think it matters that we send data from different versions.

78,876ms for bench 11 so yeah, things have changed a lot.