@lognaturel 2023-02-08
Form | Core 6.1.7 | Core 7.0.0 | Percent change
-- | -- | -- | --
bench12 | 250539 | 4949 | -98%
bench6 | 164944 | 7598 | -95%
burundi | 125001 | 2711 | -98%
bench1 | 98564 | 3278 | -97%
uganda | 79848 | 2585 | -97%
bench11 | 78876 | 2999 | -96%
shop | 19745 | 984 | -95%
bench9 | 19045 | 1414 | -93%
drc | 18991 | 1066 | -94%
haiti | 18930 | 972 | -95%
bench8 | 18243 | 847 | -95%
car | 9405 | 804 | -91%
ukraine | 8988 | 785 | -91%
turkey | 7432 | 733 | -90%
sdiprofile | 3845 | 515 | -87%
iraq | 3241 | 614 | -81%
bench4 | 2837 | 501 | -82%
bench10 | 2710 | 210 | -92%
bench2 | 2683 | 525 | -80%
bench5 | 2061 | 362 | -82%
bench3 | 728 | 180 | -75%
widgetsdev | 186 | 145 | -22%
widgets | 181 | 140 | -23%
bench7 | 137 | 110 | -20%
lagtest50 | 50 | 44 | -12%