@MartijnR 2023-02-09
GLORIOUS!! 

Here is another result:

<img width="880" alt="Screenshot 2023-02-09 at 8 46 10 AM" src="https://user-images.githubusercontent.com/627350/217829819-5486d663-453b-470e-89e8-cdfd972c7147.png">

I think I should downgrade to the cheapest droplet for this performance monitoring tool (maybe in a few days) as it looks like Enketo will no longer have any problem with that. That's a really great development for all those running big servers. (cc @jnm)