@lognaturel 2023-02-16
Thanks for transferring the perf server, @MartijnR! https://github.com/enketo/enketo-core-performance-monitor has been updated to include transformer benchmarks and use the new URL.